Add PatternPreview element #12
