+++ src/UiPatternsManagerInterface.php
@@ -10,17 +10,6 @@ use Drupal\Component\Plugin\PluginManagerInterface;
 interface UiPatternsManagerInterface extends PluginManagerInterface {
 
   /**
-   * Render patter preview.
-   *
-   * @param string $pattern_id
-   *    Pattern ID, a.k.a. the UI Patter plugin ID.
-   *
-   * @return array|\Drupal\Component\Render\MarkupInterface|string
-   *    Render array.
-   */
-  public function renderPreview($pattern_id);
-
-  /**
    * Return list of available patterns to be used as select options.
    *
    * @return array
